Rename service to protocol files.
