+++ trunk/configure
@@ -226,8 +226,8 @@ ModuleLibIncs=(${SRS_OBJS_DIR} ${LibSTRoot} ${LibSSLRoot})
 MODULE_FILES=("srs_protocol_amf0" "srs_protocol_io" "srs_rtmp_stack"
         "srs_rtmp_handshake" "srs_protocol_utility" "srs_rtmp_msg_array" "srs_protocol_stream"
         "srs_raw_avc" "srs_rtsp_stack" "srs_http_stack" "srs_protocol_kbps" "srs_protocol_json"
-        "srs_protocol_format" "srs_service_log" "srs_service_st" "srs_service_http_client" "srs_service_http_conn"
-        "srs_service_rtmp_conn" "srs_service_utility" "srs_service_conn")
+        "srs_protocol_format" "srs_protocol_log" "srs_protocol_st" "srs_protocol_http_client"
+        "srs_protocol_http_conn" "srs_protocol_rtmp_conn" "srs_protocol_conn")
 if [[ $SRS_RTC == YES ]]; then
     MODULE_FILES+=("srs_rtc_stun_stack")
     ModuleLibIncs+=(${LibSrtpRoot})
