+++ trunk/src/app/srs_app_coworkers.cpp
@@ -14,7 +14,7 @@ using namespace std;
 #include <srs_rtmp_stack.hpp>
 #include <srs_app_config.hpp>
 #include <srs_protocol_utility.hpp>
-#include <srs_service_utility.hpp>
+#include <srs_protocol_utility.hpp>
 #include <srs_kernel_utility.hpp>
 
 SrsCoWorkers* SrsCoWorkers::_instance = NULL;
