+++ trunk/src/app/srs_app_http_conn.hpp
@@ -13,7 +13,7 @@
 #include <string>
 #include <vector>
 
-#include <srs_service_http_conn.hpp>
+#include <srs_protocol_http_conn.hpp>
 #include <srs_app_reload.hpp>
 #include <srs_kernel_file.hpp>
 #include <srs_app_st.hpp>
