+++ trunk/src/app/srs_app_rtc_conn.hpp
@@ -9,7 +9,7 @@
 
 #include <srs_core.hpp>
 #include <srs_app_listener.hpp>
-#include <srs_service_st.hpp>
+#include <srs_protocol_st.hpp>
 #include <srs_kernel_utility.hpp>
 #include <srs_rtmp_stack.hpp>
 #include <srs_app_hybrid.hpp>
@@ -21,7 +21,7 @@
 #include <srs_app_rtc_queue.hpp>
 #include <srs_app_rtc_source.hpp>
 #include <srs_app_rtc_dtls.hpp>
-#include <srs_service_conn.hpp>
+#include <srs_protocol_conn.hpp>
 #include <srs_app_conn.hpp>
 #include <srs_app_async_call.hpp>
