+++ trunk/src/app/srs_app_rtc_source.hpp
@@ -17,7 +17,7 @@
 #include <map>
 
 #include <srs_app_rtc_sdp.hpp>
-#include <srs_service_st.hpp>
+#include <srs_protocol_st.hpp>
 #include <srs_app_source.hpp>
 #include <srs_kernel_rtc_rtp.hpp>
