+++ trunk/src/app/srs_app_server.hpp
@@ -18,7 +18,7 @@
 #include <srs_app_hls.hpp>
 #include <srs_app_listener.hpp>
 #include <srs_app_conn.hpp>
-#include <srs_service_st.hpp>
+#include <srs_protocol_st.hpp>
 #include <srs_app_hourglass.hpp>
 #include <srs_app_hybrid.hpp>
