+++ trunk/src/protocol/srs_protocol_http_conn.cpp
@@ -4,7 +4,7 @@
 // SPDX-License-Identifier: MIT or MulanPSL-2.0
 //
 
-#include <srs_service_http_conn.hpp>
+#include <srs_protocol_http_conn.hpp>
 
 #include <stdlib.h>
 #include <sstream>
@@ -13,12 +13,12 @@ using namespace std;
 #include <srs_kernel_log.hpp>
 #include <srs_kernel_error.hpp>
 #include <srs_protocol_stream.hpp>
-#include <srs_service_utility.hpp>
+#include <srs_protocol_utility.hpp>
 #include <srs_kernel_utility.hpp>
 #include <srs_protocol_utility.hpp>
 #include <srs_core_autofree.hpp>
 #include <srs_rtmp_stack.hpp>
-#include <srs_service_conn.hpp>
+#include <srs_protocol_conn.hpp>
 
 SrsHttpParser::SrsHttpParser()
 {
