+++ trunk/src/protocol/srs_protocol_log.cpp
@@ -4,7 +4,7 @@
 // SPDX-License-Identifier: MIT or MulanPSL-2.0
 //
 
-#include <srs_service_log.hpp>
+#include <srs_protocol_log.hpp>
 
 #include <stdarg.h>
 #include <sys/time.h>
