+++ trunk/src/utest/srs_utest_config.cpp
@@ -13,7 +13,7 @@ using namespace std;
 #include <srs_app_source.hpp>
 #include <srs_core_performance.hpp>
 #include <srs_kernel_utility.hpp>
-#include <srs_service_st.hpp>
+#include <srs_protocol_st.hpp>
 #include <srs_rtmp_stack.hpp>
 #include <srs_utest_kernel.hpp>
