+++ trunk/src/utest/srs_utest_protocol.cpp
@@ -16,7 +16,7 @@ using namespace std;
 #include <srs_app_st.hpp>
 #include <srs_protocol_amf0.hpp>
 #include <srs_rtmp_stack.hpp>
-#include <srs_service_http_conn.hpp>
+#include <srs_protocol_http_conn.hpp>
 
 MockEmptyIO::MockEmptyIO()
 {
